GH-74 GitHub review (#74)
* Start of creating review system

* Add metion reviewers to discord channel

* Create service and start creating scheduler

* Add task to delete if title conventions has problems

* Add more useful functions to utils

* More improves

* Refactor GitHubReviewUtil to handle GitHub API requests more efficiently

* Setup unit test with mock web server

* Add function to automatic create all post on forum based od github organization PR's

* Add more code

* Add database to review

* Finish review system

* improve codestyle

* Final change

* remove useless spaces

* improve test

* Follow Jakub suggestions

* Follow Karol suggestions

* Check if reviewers length is 0?

* Add option to send dm messages  if assigned as reviewer

* Add review forum id to config

* Add check if forum post is pull request

* Follow rollczi's suggestions

* Follow rollczi's suggestions

* change time to 8h

+++ src/main/java/com/eternalcode/discordapp/filter/FilterMessageEmbedController.java
@@ -14,7 +14,22 @@ public class FilterMessageEmbedController extends ListenerAdapter {
     @Override
     public void onMessageReceived(MessageReceivedEvent event) {
         event.getMessage().getEmbeds().forEach(embed -> {
-            FilterResult result = this.filterService.check(embed.getAuthor().getName(), embed.getTitle());
+            if (embed == null) {
+                return;
+            }
+
+            if (embed.getAuthor() == null) {
+                return;
+            }
+
+            String name = embed.getAuthor().getName();
+            String title = embed.getTitle();
+
+            if (name == null || title == null) {
+                return;
+            }
+
+            FilterResult result = this.filterService.check(name, title);
 
             if (!result.isPassed()) {
                 event.getMessage().delete().queue();
